+++ src/avt/Pipeline/Data/avtMIPRangeRF.C
@@ -126,7 +126,7 @@ avtMIPRangeRF::~avtMIPRangeRF()
 
 void
 avtMIPRangeRF::GetRayValue(const avtRay *ray, 
-                           unsigned char rgb[3], float depth)
+                           unsigned char rgb[3], double depth)
 {
     double curMax = -1. * DBL_MAX;
     int    maxInd = -1;
@@ -140,7 +140,7 @@ avtMIPRangeRF::GetRayValue(const avtRay *ray,
     const bool   *validSample = ray->validSample;
 
     // Only handle one variable, even if there is more.
-    const float  *sample      = ray->sample[0];
+    const double  *sample      = ray->sample[0];
 
     int depthIndex = IndexOfDepth(depth, numSamples);
